% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometries.R
\name{agb.allometry}
\alias{agb.allometry}
\title{Estimate aboveground biomass}
\usage{
agb.allometry(wsg, dbh, H)
}
\description{
from wood specific gravity (wsg), tree diameter in centimeters (dbh),
and total tree height in meters (H)
using allometries from Chave et al. (2005) for most tropical forests.
}
