% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometries.R
\name{apply.eqn1}
\alias{apply.eqn1}
\title{Calculate equivalent diameter at 1.3 m}
\usage{
apply.eqn1(d, h, b1)
}
\description{
given a diameter measurement in cm (d) taken at a nonstandard height in m(h),
and a taper parameter (b1). This is the overall best taper model
from Metcalf et el. (2009)
}
