% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stan_model_attributes.R
\name{bayesian_rsquared_production}
\alias{bayesian_rsquared_production}
\title{Get the Bayesian R-squared and its quantiles from Stan fit}
\usage{
bayesian_rsquared_production(fit, x, y, prod_model)
}
\description{
Get the Bayesian R-squared and its quantiles from Stan fit
}
