% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{binprod}
\alias{binprod}
\title{Continuous-data (point cloud) binning for production by light per crown area}
\usage{
binprod(dat, bindat)
}
\description{
Continuous-data (point cloud) binning for production by light per crown area
}
