% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{binscore}
\alias{binscore}
\title{Continuous-data (point cloud) binning for PCA values}
\usage{
binscore(dat, bindat, score_column, class_column)
}
\description{
Continuous-data (point cloud) binning for PCA values
}
