% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{cloudbin_across_years}
\alias{cloudbin_across_years}
\title{Create bins of continuous data across one or more censuses}
\usage{
cloudbin_across_years(
  dat_values,
  dat_classes,
  edges,
  mean_type = "geometric",
  n_census = 5
)
}
\description{
Create bins of continuous data across one or more censuses
}
