% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{fakebin_across_years}
\alias{fakebin_across_years}
\title{Create bins of continuous data across one or more censuses}
\usage{
fakebin_across_years(dat_values, dat_classes, edges,
  mean_type = "geometric", n_census = 5)
}
\description{
Create bins of continuous data across one or more censuses
}
