% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{insolation}
\alias{insolation}
\title{Insolation at a given latitude (approximation)}
\usage{
insolation(lat)
}
\arguments{
\item{lat}{Latitude in decimal degrees.}
}
\description{
Insolation at a given latitude (approximation)
}
