% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometries.R
\name{log_correction_factor}
\alias{log_correction_factor}
\title{Correction factor to remove bias from log-transformation}
\usage{
log_correction_factor(log_y, log_y_pred, n_pars, base = exp(1))
}
\arguments{
\item{log_y}{Log-transformed observed y-values}

\item{log_y_pred}{Log-transformed fitted y-values}

\item{n_pars}{Number of parameters in regression (used to calculate degrees of freedom)}

\item{base}{Base of the log transformation. Default to natural logarithm.}
}
\description{
Cite: Sprugel 1983 (Ecology).
}
