% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{logbin}
\alias{logbin}
\title{Logarithmic binning function}
\usage{
logbin(x, y = NULL, n)
}
\arguments{
\item{x}{Vector of values to be binned in the x dimension. Usually diameter.}

\item{y}{Vector of values to be binned in the y dimension (not required, defaults to \code{NULL}). Usually something like biomass production.}

\item{n}{Integer, number of bins}
}
\value{
A data frame with columns \code{bin_midpoint} (midpoint of bins on x-axis), \code{bin_value} (height of bins), \code{bin_count} (number of individuals in bins), \code{bin_min} (lower edge of bins on x-axis), \code{bin_max} (upper edge of bins on x-axis).
}
\description{
Logarithmic binning function
}
\seealso{
\code{\link{logbin_setedges}}
}
