% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{logseq}
\alias{logseq}
\title{Evenly spaced sequence in log space}
\usage{
logseq(x, y = NULL, l)
}
\arguments{
\item{x}{If length 1, the minimum value. If length 2, the range of values.}

\item{y}{The maximum value if not provided in x.}

\item{l}{Length of the sequence}
}
\description{
Evenly spaced sequence in log space
}
