% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{midpts}
\alias{midpts}
\title{Midpoints of a vector}
\usage{
midpts(a)
}
\arguments{
\item{a}{Numeric vector}
}
\value{
Vector with length \code{length(a) - 1}
}
\description{
Midpoints of a vector
}
