% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometries.R
\name{notaper.H}
\alias{notaper.H}
\title{Estimate total tree height (m) from diameter (cm)}
\usage{
notaper.H(dbh)
}
\description{
with uncorrected (notaper.H) or taper-corrected (taper.H) allometries parameterized from
data from the BCI plot. The function is the weibull function as described in
Feldpausch et al. 2012
}
