% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stan_model_attributes.R
\name{param_values}
\alias{param_values}
\title{Extract parameter values and their credible intervals from Stan fit}
\usage{
param_values(fit, pars)
}
\description{
Extract parameter values and their credible intervals from Stan fit
}
