% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pct_light_captured}
\alias{pct_light_captured}
\title{Proportion light captured by a crown}
\usage{
pct_light_captured(depth, k)
}
\arguments{
\item{depth}{Depths of crowns in meters (vector)}

\item{k}{Light extinction coefficient (>0)}
}
\description{
Proportion light captured by a crown
}
