% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_relationships.R
\name{pdf_3part}
\alias{pdf_3part}
\title{PDF for density model 3 part}
\usage{
pdf_3part(x, xmin, alpha_low, alpha_mid, alpha_high, tau_low, tau_high)
}
\description{
PDF for density model 3 part
}
