% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bins_and_fits.R
\name{plot_prod_fixed}
\alias{plot_prod_fixed}
\title{Another variant of production plotting function}
\usage{
plot_prod_fixed(year_to_plot = 1995, fg_names = c("fg1", "fg2", "fg3",
  "fg4", "fg5", "all"), model_fit = 1, x_limits, x_breaks = c(1, 3, 10,
  30, 100, 300), y_limits, y_labels, y_breaks, fill_names = c("#BFE046",
  "#267038", "#27408b", "#87Cefa", "ivory"), fill_names0 = c("#BFE046",
  "#267038", "#27408b", "#87Cefa", "gray"), x_name = "Diameter (cm)",
  y_name = expression(paste("Growth (kg y"^-1, ")")), average = "mean",
  error_quantiles = c("ci_min", "ci_max"), error_bar_width = 0.03,
  dodge_width = 0.03, dodge_errorbar = TRUE, geom_size = 4,
  obsdat = obs_indivprod, preddat = fitted_indivprod)
}
\description{
Another variant of production plotting function
}
