% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bins_and_fits.R
\name{plot_prod_withrawdata}
\alias{plot_prod_withrawdata}
\title{Plot hexbin plot of production and/or fitted production functions}
\usage{
plot_prod_withrawdata(year_to_plot = 1995, fg_names = c("fg1", "fg2",
  "fg3", "fg4", "fg5", "unclassified"), full_names = c("Fast", "Slow",
  "Pioneer", "Breeder", "Medium", "Unclassified"),
  func_names = c("power law", "2-segment power law"), x_limits = c(1,
  300), x_breaks = c(1, 10, 100), y_limits, y_breaks,
  x_name = "Diameter (cm)",
  y_name = expression(paste("Growth (kg yr"^-1, ")")),
  line_types = c("dashed", "solid"), aspect_ratio = 0.75,
  hex_scale = ggplot2::scale_fill_gradientn(colours =
  colorRampPalette(rev(RColorBrewer::brewer.pal(9, "RdYlBu")), bias =
  1)(50), trans = "log", name = "Individuals", breaks = c(1, 10, 100, 1000,
  10000), labels = c(1, 10, 100, 1000, 10000), limits = c(1, 10000)),
  obsdat = raw_prod, preddat = fitted_indivprod, plot_abline = TRUE,
  plot_fits = FALSE)
}
\description{
Arguments not described. Sorry. See \code{\link{plot_prod}} for similar arguments.
}
