% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bins_and_fits.R
\name{plot_totalprod}
\alias{plot_totalprod}
\title{Plot single model fit with multiple functional groups for total production}
\usage{
plot_totalprod(
  year_to_plot = 1995,
  fg_names = c("fg1", "fg2", "fg3", "fg4", "fg5", "all"),
  model_fit_density = 1,
  model_fit_production = 1,
  x_limits,
  x_breaks = c(1, 3, 10, 30, 100, 300),
  y_limits = c(0.03, 100),
  y_breaks = c(0.01, 0.1, 1, 10, 100, 1000),
  y_labels,
  fill_names = c("black", "#BFE046", "#267038", "#27408b", "#87Cefa", "gray87"),
  color_names = c("black", "#BFE046", "#267038", "#27408b", "#87Cefa", "gray"),
  x_name = "Diameter (cm)",
  y_name = expression(paste("Production (kg ha"^-1, " cm"^-1, " y"^-1, ")")),
  geom_size = 4,
  obsdat = obs_totalprod,
  preddat = fitted_totalprod,
  plot_abline = TRUE,
  abline_slope = 0,
  abline_intercept = 2
)
}
\arguments{
\item{year_to_plot}{Default 1995, currently no other years have model fits.}

\item{fg_names}{Vector of functional group names to plot. Defaults to plot all five plus the total: \code{c('fg1','fg2','fg3','fg4','fg5','all')}}

\item{model_fit_density}{Which density model to plot. Defaults to 1 (one-segment power law) but 2 was also fit and can be plotted.}

\item{model_fit_production}{Which production model to plot. Defaults to 1 (one-segment power law) but 2 was also fit and can be plotted.}

\item{x_limits}{2-length vector of minimum and maximum x-value to plot.}

\item{x_breaks}{Tick values to display on x-axis.}

\item{y_limits}{2-length vector of minimum and maximum y-value to plot.}

\item{y_breaks}{Tick values to display on y-axis.}

\item{y_labels}{Labels to show at the tick values on the y-axis (in case you want different notation).}

\item{fill_names}{Vector of fill colors. Must be same length as \code{fg_names}.}

\item{color_names}{Vector of line colors. Must be same length as \code{fg_names}.}

\item{x_name}{Title of x-axis.}

\item{y_name}{Title of y-axis.}

\item{geom_size}{Size of points to plot. Default is 4.}

\item{plot_abline}{Whether to include reference line, default TRUE}

\item{abline_slope}{Slope of the optional reference line, default 0}

\item{abline_intercept}{Intercept of the optional reference line, default 2}

\item{obs_dens}{Name of R object with observed density values.}

\item{pred_dens}{Name of R object with fitted density values.}
}
\description{
Specify both density and production type
}
