% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_relationships.R
\name{powerlaw_hinge_log}
\alias{powerlaw_hinge_log}
\title{Function for production model 2 parts (hinged)}
\usage{
powerlaw_hinge_log(x, x0, beta0, beta1_low, beta1_high, delta)
}
\description{
Function for production model 2 parts (hinged)
}
