% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometries.R
\name{taper_correct_dbh}
\alias{taper_correct_dbh}
\title{Calculate the taper-corrected DBH for all individuals in one data frame.}
\usage{
taper_correct_dbh(dat)
}
\description{
Calculate the taper-corrected DBH for all individuals in one data frame.
}
