% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_plant}
\alias{theme_plant}
\alias{theme_plant_small}
\alias{theme_facet}
\alias{theme_facet2}
\alias{theme_no_y}
\alias{theme_no_x}
\title{Forestlight plotting themes}
\usage{
theme_plant()

theme_plant_small()

theme_facet()

theme_facet2()

theme_no_y()

theme_no_x()
}
\value{
Returns a ggplot2 theme that can be added to an existing plot. Note that parentheses must be used.
}
\description{
Forestlight plotting themes
}
\examples{
ggplot(data, aes(x, y)) + geom_line() + theme_plant()

}
